requirement.txt
GitPython
python-docx
openai
tqdm